refactor: add disable/enable views
